refactor: extract centered_rect_with_size to shared utility
- Created src/ui/utils.rs with centered_rect_with_size function
- Removed 9 duplicate implementations across modal files
- Updated all modal files to use the shared utility
- Net reduction: -152 LOC (61 added, 213 removed)

This eliminates code duplication and makes modal centering logic
easier to maintain and modify in the future.
