+++ src/ui/mod.rs
@@ -4,5 +4,7 @@ pub mod menu;
 pub mod modal;
 pub mod panel_rendering;
 pub mod status_bar;
+pub mod utils;
 
 pub use layout::render_layout_with_accordion;
+pub use utils::centered_rect_with_size;
