+++ src/ui/modal/confirm.rs
@@ -11,6 +11,7 @@ use ratatui::{
 use super::{Modal, ModalResult};
 use crate::i18n;
 use crate::theme::Theme;
+use crate::ui::centered_rect_with_size;
 
 /// Confirmation modal window (Yes/No)
 #[derive(Debug)]
@@ -62,33 +63,6 @@ impl ConfirmModal {
         let max_width = (screen_width as f32 * 0.75) as u16;
         total_width.max(20).min(max_width).min(screen_width)
     }
-
-    /// Create a centered rectangle with fixed size
-    fn centered_rect_with_size(width: u16, height: u16, r: Rect) -> Rect {
-        use ratatui::layout::{Constraint, Direction, Layout};
-
-        // Calculate margins
-        let horizontal_margin = r.width.saturating_sub(width) / 2;
-        let vertical_margin = r.height.saturating_sub(height) / 2;
-
-        let vertical_layout = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([
-                Constraint::Length(vertical_margin),
-                Constraint::Length(height),
-                Constraint::Length(vertical_margin),
-            ])
-            .split(r);
-
-        Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([
-                Constraint::Length(horizontal_margin),
-                Constraint::Length(width),
-                Constraint::Length(horizontal_margin),
-            ])
-            .split(vertical_layout[1])[1]
-    }
 }
 
 impl Modal for ConfirmModal {
@@ -104,7 +78,7 @@ impl Modal for ConfirmModal {
         let modal_width = self.calculate_modal_width(area.width);
 
         // Create centered area with calculated dimensions
-        let modal_area = Self::centered_rect_with_size(modal_width, modal_height, area);
+        let modal_area = centered_rect_with_size(modal_width, modal_height, area);
 
         // Clear the area
         Clear.render(modal_area, buf);
