+++ src/ui/modal/input.rs
@@ -11,6 +11,7 @@ use ratatui::{
 use super::{Modal, ModalResult};
 use crate::i18n;
 use crate::theme::Theme;
+use crate::ui::centered_rect_with_size;
 
 /// Focus area in the modal
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -115,33 +116,6 @@ impl InputModal {
 
         (width, height)
     }
-
-    /// Create a centered rectangle with fixed size
-    fn centered_rect_with_size(width: u16, height: u16, r: Rect) -> Rect {
-        use ratatui::layout::{Constraint, Direction, Layout};
-
-        // Calculate margins
-        let horizontal_margin = r.width.saturating_sub(width) / 2;
-        let vertical_margin = r.height.saturating_sub(height) / 2;
-
-        let vertical_layout = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([
-                Constraint::Length(vertical_margin),
-                Constraint::Length(height),
-                Constraint::Length(vertical_margin),
-            ])
-            .split(r);
-
-        Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([
-                Constraint::Length(horizontal_margin),
-                Constraint::Length(width),
-                Constraint::Length(horizontal_margin),
-            ])
-            .split(vertical_layout[1])[1]
-    }
 }
 
 impl Modal for InputModal {
@@ -152,7 +126,7 @@ impl Modal for InputModal {
         let (modal_width, modal_height) = self.calculate_modal_size(area.width, area.height);
 
         // Create centered area
-        let modal_area = Self::centered_rect_with_size(modal_width, modal_height, area);
+        let modal_area = centered_rect_with_size(modal_width, modal_height, area);
 
         // Clear the area
         Clear.render(modal_area, buf);
