+++ src/ui/modal/overwrite.rs
@@ -10,6 +10,7 @@ use ratatui::{
 
 use super::{Modal, ModalResult};
 use crate::theme::Theme;
+use crate::ui::centered_rect_with_size;
 
 /// File overwrite action options
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -104,33 +105,6 @@ impl OverwriteModal {
         let max_width = (screen_width as f32 * 0.75) as u16;
         total_width.max(20).min(max_width).min(screen_width)
     }
-
-    /// Create a centered rectangle with fixed size
-    fn centered_rect_with_size(width: u16, height: u16, r: Rect) -> Rect {
-        use ratatui::layout::{Constraint, Direction, Layout};
-
-        // Calculate margins
-        let horizontal_margin = r.width.saturating_sub(width) / 2;
-        let vertical_margin = r.height.saturating_sub(height) / 2;
-
-        let vertical_layout = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([
-                Constraint::Length(vertical_margin),
-                Constraint::Length(height),
-                Constraint::Length(vertical_margin),
-            ])
-            .split(r);
-
-        Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([
-                Constraint::Length(horizontal_margin),
-                Constraint::Length(width),
-                Constraint::Length(horizontal_margin),
-            ])
-            .split(vertical_layout[1])[1]
-    }
 }
 
 impl Modal for OverwriteModal {
@@ -145,7 +119,7 @@ impl Modal for OverwriteModal {
         let modal_height = 8;
 
         // Create centered area
-        let modal_area = Self::centered_rect_with_size(modal_width, modal_height, area);
+        let modal_area = centered_rect_with_size(modal_width, modal_height, area);
         Clear.render(modal_area, buf);
 
         // Create block with inverted colors
